+++ R/fasstr-package.R
@@ -14,7 +14,6 @@
 #' 
 #' tidyhydat information:
 #' 
-#' 
 #' To use the station_number argument of the fasstr functions, please download the latest version of HYDAT using the function:
 #'  \itemize{
 #'   \item \code{tidyhydat::download_hydat()}
@@ -25,9 +24,9 @@
 #'  \itemize{
 #'    \item Please see this description of the database: goo.gl/H3NXJQ
 #'    \item This page is landing page for technical description of HYDAT:
-#'    \url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/}
+#'    \url{https://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/}
 #'    \item This page links to a document that outlines database table definitions:
-#'    \url{http://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/HYDAT_Definition_EN.pdf}
+#'    \url{https://collaboration.cmc.ec.gc.ca/cmc/hydrometrics/www/HYDAT_Definition_EN.pdf}
 #'    }
 #'
 #' @keywords internal