+++ R/plot_monthly_cumulative_stats.R
@@ -13,10 +13,10 @@
 #' @title Plot cumulative monthly flow statistics
 #'
 #' @description Plot the monthly cumulative mean, median, maximum, minimum, and 5, 25, 75, 95th percentiles for each month of the year 
-#'    from a streamflow dataset. Plots the statistics from all monthly cumulative values from all years, unless specified. 
-#'    Data calculated using calc_monthly_cumulative_stats() function. Can plot individual years for comparison using the 
+#'    from a daily streamflow data set. Calculates statistics from all values from complete years, unless specified. 
+#'    Data calculated using \code{calc_monthly_cumulative_stats()} function. Can plot individual years for comparison using the 
 #'    add_year argument. Defaults to volumetric cumulative flows, can use \code{use_yield} and \code{basin_area} to convert to 
-#'    runoff yield.
+#'    water yield. Returns a list of plots.
 #'    
 #' @inheritParams calc_monthly_cumulative_stats
 #' @inheritParams plot_daily_cumulative_stats
@@ -199,7 +199,7 @@ plot_monthly_cumulative_stats <- function(data,
                                                           long = ggplot2::unit(.2, "cm"))} +
          ggplot2::xlab("Month")+
          ggplot2::scale_x_discrete(expand = c(0.01,0.01)) +
-         {if(!use_yield) ggplot2::ylab(expression(Cumulative~Volume~(m^3)))} +
+         {if (!use_yield) ggplot2::ylab("Cumulative Volume (cubic metres)")} +
          {if(use_yield) ggplot2::ylab("Cumulative Yield (mm)")} +
          ggplot2::theme_bw() +
          ggplot2::labs(color = 'Monthly Statistics') +