+++ R/zzz.R
@@ -92,7 +92,9 @@
           "Info",
           "Month_Mean",
           "Annual_Mean",
-          "miss_plots")
+          "miss_plots",
+          "MAX_VALUE",
+          "Return_P")
       )
     invisible()
   }