+++ inst/doc/fasstr_under_the_hood.R
@@ -1,8 +1,7 @@
 ## ----options, include=FALSE---------------------------------------------------
-knitr::opts_chunk$set(eval = nzchar(Sys.getenv("hydat_eval")))
-
-## ---- include=FALSE-----------------------------------------------------------
+knitr::opts_chunk$set(eval = nzchar(Sys.getenv('hydat_eval')))
 library(fasstr)
+
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  # Check if data is provided and import it
@@ -29,20 +28,19 @@ library(fasstr)
 #  flow_data <- dplyr::mutate(flow_data, Yield_mm = Value * 86400 / (Basin_Area_sqkm_temp * 1000))
 #  
 #  # Return the original names of the Date and Value columns
-#  names(flow_data)[names(flow_data) == "Value"] <- as.character(substitute(values))
-#  names(flow_data)[names(flow_data) == "STATION_NUMBER"] <- as.character(substitute(groups))
+#  names(flow_data)[names(flow_data) == 'Value'] <- as.character(substitute(values))
+#  names(flow_data)[names(flow_data) == 'STATION_NUMBER'] <- as.character(substitute(groups))
 #  
 #  
 #  ## Reformat to original names and groups
 #  ## -------------------------------------
 #  
 #  # Return columns to original order plus new column
-#  if("Yield_mm" %in% orig_cols){
+#  if('Yield_mm' %in% orig_cols){
 #    flow_data <-  flow_data[, c(orig_cols)]
 #  } else {
-#    flow_data <-  flow_data[, c(orig_cols, paste("Yield_mm"))]
+#    flow_data <-  flow_data[, c(orig_cols, paste('Yield_mm'))]
 #  }
-#  
 #  
 #  dplyr::as_tibble(flow_data)
 
@@ -53,16 +51,16 @@ library(fasstr)
 #  
 #  # Add rolling means to end of dataframe
 #  flow_data <- add_rolling_means(data = flow_data, roll_days = roll_days, roll_align = roll_align)
-#  colnames(flow_data)[ncol(flow_data)] <- "RollingValue"
+#  colnames(flow_data)[ncol(flow_data)] <- 'RollingValue'
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  # Filter for the selected year (remove excluded years after)
-#  flow_data <- dplyr::filter(flow_data, AnalysisYear >= start_year & AnalysisYear <= end_year)
+#  flow_data <- dplyr::filter(flow_data, WaterYear >= start_year & WaterYear <= end_year)
 #  flow_data <- dplyr::filter(flow_data, Month %in% months)
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  # Calculate basic stats
-#  annual_stats <-   dplyr::summarize(dplyr::group_by(flow_data, STATION_NUMBER, AnalysisYear),
+#  annual_stats <-   dplyr::summarize(dplyr::group_by(flow_data, STATION_NUMBER, WaterYear),
 #                                     Mean = mean(RollingValue, na.rm = ignore_missing),
 #                                     Median = stats::median(RollingValue, na.rm = ignore_missing),
 #                                     Maximum = max (RollingValue, na.rm = ignore_missing),
@@ -78,13 +76,13 @@ library(fasstr)
 #  if(!all(is.na(percentiles))) {
 #    for (ptile in percentiles) {
 #      # Calculate percentiles
-#      annual_stats_ptile <- dplyr::summarise(dplyr::group_by(flow_data, STATION_NUMBER, AnalysisYear),
+#      annual_stats_ptile <- dplyr::summarise(dplyr::group_by(flow_data, STATION_NUMBER, WaterYear),
 #                                             Percentile = stats::quantile(RollingValue, ptile / 100, na.rm = TRUE))
 #      annual_stats_ptile <- dplyr::ungroup(annual_stats_ptile)
-#      names(annual_stats_ptile)[names(annual_stats_ptile) == "Percentile"] <- paste0("P", ptile)
+#      names(annual_stats_ptile)[names(annual_stats_ptile) == 'Percentile'] <- paste0('P', ptile)
 #  
 #      # Merge with stats
-#      annual_stats <- merge(annual_stats, annual_stats_ptile, by = c("STATION_NUMBER", "AnalysisYear"))
+#      annual_stats <- merge(annual_stats, annual_stats_ptile, by = c('STATION_NUMBER', 'WaterYear'))
 #  
 #      # Remove percentile if mean is NA (workaround for na.rm=FALSE in quantile)
 #      annual_stats[, ncol(annual_stats)] <- ifelse(is.na(annual_stats$Mean), NA, annual_stats[, ncol(annual_stats)])
@@ -93,7 +91,7 @@ library(fasstr)
 
 ## ---- eval=FALSE--------------------------------------------------------------
 #  # Rename year column
-#  annual_stats <- dplyr::rename(annual_stats, Year = AnalysisYear)
+#  annual_stats <- dplyr::rename(annual_stats, Year = WaterYear)
 #  
 #  # Remove selected excluded years
 #  annual_stats[annual_stats$Year %in% exclude_years, -(1:2)] <- NA
@@ -119,17 +117,16 @@ library(fasstr)
 #  
 #  # Recheck if station_number/grouping was in original data and rename or remove as necessary
 #  if(as.character(substitute(groups)) %in% orig_cols) {
-#    names(annual_stats)[names(annual_stats) == "STATION_NUMBER"] <- as.character(substitute(groups))
+#    names(annual_stats)[names(annual_stats) == 'STATION_NUMBER'] <- as.character(substitute(groups))
 #  } else {
 #    annual_stats <- dplyr::select(annual_stats, -STATION_NUMBER)
 #  }
-#  
 #  
 #  dplyr::as_tibble(annual_stats)
 
 ## ---- fig.height = 3, fig.width = 7, comment=NA-------------------------------
 # Calculate the statistics
-annual_stats <- calc_annual_stats(station_number = c("08NM116", "08NM240"),
+annual_stats <- calc_annual_stats(station_number = c('08NM116', '08NM240'),
                                   start_year = 1985, end_year = 2015)
 
 # Wrangle statistics for plotting
@@ -148,7 +145,7 @@ tidy_plots <- dplyr::mutate(tidy_plots,
                                                  ggplot2::theme(plot.title = ggplot2::element_text(hjust = 0.5)) +
                                                  ggplot2::geom_line(alpha = 0.5, na.rm = TRUE) +
                                                  ggplot2::geom_point(na.rm = TRUE) +
-                                                 ggplot2::ylab("Discharge (cms)")
+                                                 ggplot2::ylab('Discharge (cms)')
 
                             ))
 
@@ -156,9 +153,9 @@ tidy_plots <- dplyr::mutate(tidy_plots,
 # Create a list of named plots extracted from the tibble
 plots <- tidy_plots$plot
 if (nrow(tidy_plots) == 1) {
-  names(plots) <- "Annual_Statistics"
+  names(plots) <- 'Annual_Statistics'
 } else {
-  names(plots) <- paste0(tidy_plots$STATION_NUMBER, "_Annual_Statistics")
+  names(plots) <- paste0(tidy_plots$STATION_NUMBER, '_Annual_Statistics')
 }
 
 # Return the plots