+++ NEWS.md
@@ -1,3 +1,20 @@
+fasstr 0.3.2
+=========================
+
+Updated: 06 Nov 2020
+
+#### Updates:
+
+  * Added calc_annual_peaks() function to calculate both annual n-day min. and max. flows with dates of occurrence.
+  * Added error in volume frequency analysis when providing zero or negative flow values to the logPIII dist.
+  * Updated volume frequency analysis documentation.
+  * Updated all vignettes for package updates, corrections, and issues.
+  * Added new 'Get Started' vignette
+  * Updated Cheatsheet
+  * changed yield terminology to water yield from runoff yield
+  * changed plot axis titles with expression()'s (i.e. m3/s --> cms)
+  
+
 fasstr 0.3.1
 =========================